fix(rust): added check for max message length
