+++ examples/rust/mitm_node/src/tcp_interceptor/mod.rs
@@ -1,6 +1,5 @@
 mod registry;
 mod transport;
-pub mod utils;
 
 pub use registry::*;
 pub use transport::*;
